+++ libs/features/personalization/add-fragment-link-headers.js
@@ -1,0 +1,13 @@
+export default async function addFragmentLinkHeaders(fragment, a) {
+  fragment.dataset.manifestId = a.dataset.manifestId;
+  const manifestPrefix = a.dataset.manifestId.split(' ')[0];
+  if (!manifestPrefix.includes('.json')) {
+    let daaLh = a.getAttribute('daa-lh');
+    if (daaLh) {
+      daaLh = `${manifestPrefix} ${daaLh}`;
+    } else {
+      daaLh = manifestPrefix;
+    }
+    fragment.setAttribute('daa-lh', daaLh);
+  }
+}
